+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Changed
 
 * Upgraded hyper to 0.14 (and therefore implicitly to tokio 1.0). [#19](https://github.com/boxdot/letterboxd-rs/pull/19)
+* `letterboxd::Error` is `Send` and `Sync` [#21](https://github.com/boxdot/letterboxd-rs/pull/21)